Update button sensitivity when an image is opened or closed.
When the first image is opened, the 'iter-changed' signal is
emitted, this causes the main-window to toggle the sensitivity of
the next, previous and play buttons. Since 'next', 'previous' and
'play slideshow' do not make sense when you've opened a single
image, these 3 buttons are marked insensitive then. Opening the
rest of the folder doesn't move the iterator, and thus doesn't
trigger additional 'iter-changed' signals. Keeping the buttons
disabled even though several other images have been opened aswell.

Calling the 'update_buttons' function immediately after opening
new images or a folder of images solves this discrepancy.

For consistency, 'update_buttons' is called from the function
closing the images aswell. This makes sure the zoom- and rotate-
buttons are disabled when no image is opened.