Chain up to parent in dispose() and finalize()
For this reason, rstto_image_viewer_dispose() must be changed to
rstto_image_viewer_finalize() otherwise data are accessed after they
have been released.

Probably some or all of the dispose() functions could also be changed to
finalize() in the code, but it doesn't seem necessary at this point so
it's better to limit this commit to that.