Transfer contents from dispose() to finalize()
This will prevent any use after free of private structures, which are
typically things to be released in finalize() and not in dispose().

If this should ever prevent a reference counting cycle from being
broken, which is unlikely, it will still be possible to restore object
releases in dispose() on a case-by-case basis.

Completes 68b25dffc946abd9f7cb4c93e1aad46292a86156.