Remove the shadow of the toolbar buttons
With some themes, including Adwaita, this shadow can drool on the image
viewer when a button is pressed. More importantly, a single button hover
triggers a redraw of the image viewer which, although limited to the
size of the shadow, can make the toolbar difficult to use with large
images.

Fixes #51, partially addresses #64.