Use integers for the image rendering rectangle
This allows these values to be used directly for drawing area
restriction in functions like `gdk_window_invalidate_*()`, instead of
rounding up or down to ensure that the restriction is fully applied.

A scale adjustment is also necessary at the time of painting so that
the painted image is strictly included in the rendering rectangle, and
the restriction applies properly.

A typical example where this adjustment is crucial is when excluding
the image for redrawing the background only (implemented in the
following commit).