Fix file creation monitoring
Created files are filtered on their mime type when added through
`rstto_image_list_add_file()`. But in general, when
`G_FILE_MONITOR_EVENT_CREATED` is emitted, the file can still be empty.
In this case, its mime type is `inode/x-empty` and it is therefore not
added to the list, when it should perhaps be.

This is a case where `G_FILE_MONITOR_EVENT_CHANGES_DONE_HINT` seems
useful: it is emitted when the file creation is finished, whether it is
really a creation or a move. So we can wait for this event to add the
file to the list.

This commit also removes the code under `G_FILE_MONITOR_EVENT_MOVED`,
which is never emitted since the flag `G_FILE_MONITOR_SEND_MOVED` is
not used.