Optimize image viewer redraw: Reuse source surface
Except in the case of an animated image, where we are always juggling
with two frames, and once the creation of the source surface from the
original image is over, we should only have in memory one "copy" of
this image (in the broad sense, i.e. in one format or another).

This is the first objective of this commit. It allows, once an image
has been displayed for the first time, to keep a constant memory
occupation for any other operation requiring a redrawing of this same
image (in particular zoom, rotation, change of window size).

This also has the appreciable side effect of improving - more or less
significantly depending on the case - the performance during these
operations, since the manipulations around the source surface prior to
redrawing are only done once.

Partially addresses #64, see !26 for more details.