Optimize image viewer redraw: Disable overlay scrolling
As soon as the image is a bit big, the practicality of the
overlay does not justify the cost and the side effects that go with it:
the scrollbars actually become almost unusable if the image is big
enough.

The policy of automatically displaying scrollbars is kept in windowed
mode, but they are hidden in fullscreen mode, where one usually wants
to have as few extraneous elements as possible in the image (which one
can always drag with the left click if needed).

Partially addresses #64.