Allow to choose the default scale
This adds a radio menu bound to a new setting "default-zoom", which
applies to images displayed for the first time. To the current behavior
called "smart", which remains the default, are added "fit-to-view" and
"real-size".

If this setting is changed when an image list is already loaded, the
scale is reset for all images. This is probably what is expected in
general, and typically when switching from windowed to fullscreen mode,
using "smart" or "real-size" for the former, and "fit-to-view" for the
latter.

See !29 for more details.