Take into account the device scale
This prevents the image and the thumbnails from being scaled with the
rest of the window when a device scale is used. Thus the image is
displayed by default in real size (if the size of the view allows it)
and the quality of the thumbnails is not altered. It is then up to the
user to adjust their size via the settings.

Fixes #47, see !30 for more details.