Properly load directories via drag and drop
Completes f228cebd3977431367bf124466a981f6b8852f43.

This is a minimal fix that will be outdated in a future redesign work to
address #20, but it will provide a reproducible level of functionality
for non-regression testing.

Also fixes a memory leak already fixed twice elsewhere:
* 1939c005b62c56034c152cba70168f11dac23080
* 1b55de2d59a9ad49a8dde9824ba76c86e38c92f6
, due to a piece of code more or less well reproduced in multiple
places. This too will be redesigned in the upcoming work mentioned
above.

Related to #67.