Optimize icon bar drawing
This simply restricts item drawing to the drawing area (which is often
the visible area), which decreases the number of calls to
`rstto_icon_bar_paint_item()` from O(n) to O(1).

The cairo context already restricts drawing to this area, but this also
saves useless calls to `rstto_file_get_thumbnail()`, which contains
potential disk access and Tumbler invocations. So this can significantly
improve performance for large item numbers.

See !31 for more details.