Optimize directory loading: Directory scanning
This first step rewrites directory scanning using async GIO APIs, which
simplifies and clarifies the procedure. Progressive loading is removed
since it did not prevent the GUI from freezing and will no longer be
necessary at the end of this series of optimizations.

The main optimization here is to no longer check for each created
RsttoFile if it is already open, which decreases the complexity of a
file creation from O(n) to O(1), and that of directory scanning from
O(n^2) to O(n).

The only file that might be duplicated in the current implementation of
adding a directory is handled separately, in at most n calls to
`g_file_equal()`.

The only remaining place where checking for an already open file is
needed is in `cb_file_monitor_changed()`, so the code for this check is
moved there.

This also allows to go from O(n^2) to O(n) complexity when closing
Ristretto, since a call to `g_list_remove()` was made when each
RsttoFile was destroyed.

Error handling is left for a later refactoring work, since it was not
already done in this part and remains to be done elsewhere.

See !31 for more details.