Optimize directory loading: Adding files to the list
* Any unnecessary operations are removed when adding a file to the image
  list, so that it ultimately boils down to a call to `g_list_prepend()`.
* The list is sorted only once, which decreases the complexity by an
  order of magnitude (it should typically go from O(n^2 log(n)) to
  O(n log(n))).
* Image filtering requiring disk access is done at the last moment, when
  the thumbnails are displayed, and only for those in the visible part
  of the icon bar, which decreases the complexity from O(n) to O(1).

This last point, together with the optimization of the thumbnail bar
drawing in a previous commit, also greatly improves the smoothness of
its scrolling for large image numbers.

This also fixes the lack of re-selection of the current image in the
icon bar when changing the sort order.

See !31 for more details.