flatpak: Add generic manifest and basic scripts
This first version of the manifest embeds Xfconf as a dependency but
only uses it on the client side, assuming the existence of a D-Bus
Xfconf service on the host.

This is obviously not satisfactory, both because of the external
dependency it adds, and because of the sharing of configuration data
with the outside. But this commit allows to keep a "standard version" of
the manifest, before adding a patch to fix these problems.