flatpak: Adaptations to run our own Xfconf service
A patch is added to allow the embedded Xfconf dependency to be used on
both the client and server sides. It renames the D-Bus service from
`org.xfce.Xfconf` to `org.xfce.ristretto.Xfconf`, so it is unique and in
particular cannot be confused with the one of an Xfce session, or the
one embedded in the flatpak of another Xfce application.

Since the service is owned by Ristretto, it stores the configuration
data in its space, under
`~/.var/app/org.xfce.ristretto/config/xfce4/xfconf`.