Directory loading: Ensure GVfs location support
Using asynchronous methods both to get the directory enumerator and the
file info is needed in general, as well as specifying "standard::name"
in order to be able to use `g_file_enumerator_get_child()` afterwards.
See https://gitlab.gnome.org/GNOME/glib/-/issues/2507.

Error handling is left for a later refactoring work, since it was not
already done in this part and remains to be done elsewhere.

This fixes a regression introduced in
db21453277ec70dea4b79ae019bbd32d72e243bc, and allows to add GVfs
location support for the Flatpak format.

Related to !31.