Directory loading: Fix improper use of GFileEnumerator APIs
Using asynchronous methods both to get the directory enumerator and the
file info is needed in general, as well as specifying "standard::name"
in order to be able to use `g_file_enumerator_get_child()` afterwards.
See https://gitlab.gnome.org/GNOME/glib/-/issues/2507.

This fixes in particular a regression introduced in
db21453277ec70dea4b79ae019bbd32d72e243bc about GVfs locations, and
allows to add GVfs location support for the Flatpak format.

Error handling is left for a later refactoring work, since it was not
already done in this part and remains to be done elsewhere.

Fixes #68, related to !31.