Directory loading: Allow to load a directory via the cli
More generally, this allows to load a directory without specifying a
reference file to be loaded first. But in practice, in the current state
of the code, the only place where such a thing occurs is when specifying
a directory on the command line.

This forces the directory loading procedure to be split into a few more
pieces, so that the image has a chance to be displayed before the
blocking part of the loading.

Error handling is left for a later refactoring work, since it was not
already done in this part and remains to be done elsewhere.

Fixes a regression introduced in
dafcfab70b2b237a5a2688764986eea3dc2de756, related to !31.