Filter deleted recent files on opening
This does not completely prevent having a deleted file in the recent
menu, since it is generated only once on opening. But it limits the risk
and ensures that the cleanup is done at least every session.

This commit is a preliminary to a refactoring of the file opening
procedure. GLib goes to 2.44 because of `g_strv_contains()`.

See !33 for more details.