Rework the file opening procedure into a single function
Rather than having four more or less similar ones for the open dialog,
the recent menu, drag and drop and the command line.

In particular:
* Files opened explicitly (not implicitly via opening a directory) are
  added to the list of recent files regardless of how they were opened.
* Filtering on mime type is removed and is now only located in
  `image_list.c`. As a consequence, when opening a file via the command
  line, filtering is only content-based (if libmagic is present).

Error handling is left for the next commit, since it was not
already done in this part and remains to be done elsewhere.

See !33 for more details.