Directory loading: Filter files using the app filter
Now that the file opening procedure is unique and no longer includes
additional filtering, the correct filter to use for directory loading is
the application filter.

This is consistent with the open dialog on the one hand, and on the
other hand as close as possible to the final filtering, which takes
place at the last moment with additional disk access, when the
thumbnails are displayed.

See !33 for more details.