Increase image loading buffer size
This size was reduced about ten years ago in
e49a5262732e31cdf4cba348212f099c4a620613, with the argument that too
large a size caused problems loading JPEG images. However, no link was
given as a reference, and the issues of this type found on the
gdk-pixbuf bug tracker have been closed due to lack of activity.

So it seems to be safe to assume that the problem has since been fixed,
perhaps outside of gdk-pixbuf.