Use the generic pixbuf loader for RAW image files
The pixbuf loader provided by libopenraw is too broken to be used, see
https://gitlab.freedesktop.org/libopenraw/libopenraw/-/issues/7.

In general, the generic loader will just retrieve the natively
recognized low quality image embedded in the RAW file (typically JPEG or
TIFF). However, for some images, it may use some features offered by the
library to display "something" (probably of poor quality in general). So
it's better to keep the use of the library, rather than disabling it
completely with `gdk_pixbuf_format_set_disabled()`.

By the way, what is implemented in this commit is generic and could be
used in the future for other image formats, to which the above remark
would also apply.

Regarding the "image/x-canon-cr2" mime type recognition added in
6d43825421d77c3fefca3df9202d5e5ae9a5b89c, it is better to require the
user to install libopenraw, and let it communicate with gdk-pixbuf to
feed the filter via `gtk_file_filter_add_pixbuf_formats()`. Especially
since this library is also needed by Tumbler to provide the thumbnails.

GLib goes to 2.44 because of `g_strv_contains()`.