Thumbnailer: Do not empty an unprocessed file list
When emptying the file list to be processed on the "finished" signal,
you may find yourself in the situation where this signal is sent for a
previous list, while a new one has already been filled. This has the
effect of emptying the current list when it has not yet been processed.

To avoid this, it is better to empty the list only before refilling it
and when destroying the thumbnailer. Also, the corresponding requests
should be cancelled at this time.