Optimize sorting by content type
This stores the content type obtained at a lower cost during a directory
loading, so that it can be used for an eventual sorting of the image
list on this criterion, without additional disk access.

The content type is then refined on a case-by-case basis when displaying
the images, either in the view or in the thumbnail bar (see previous
commit). This refinement is now moved to the filtering function.

Related to !31, completes dafcfab70b2b237a5a2688764986eea3dc2de756.