Thumbnailer: Properly set the number of visible items
It actually corresponded to the number of items drawn, which is not
always the same thing. In addition, it remained at zero when the icon
bar was hidden, which made it impossible, for example, to send a request
to update the window icon.

The way of retrieving information and related objects is redesigned on
this occasion.

Related to !31.