Avoid multi-threading issue with X11
From 55afcd6e9673b324ef595d1981d15da8ea70b15d, the image loader sends
its signals in multi-threaded context. Concerning the "size-prepared"
signal, when the rendering quality limitation option is active, the
screen size must be retrieved. This leads GDK to use X11 specific APIs
when this windowing system is used.

In order not to have to distinguish between windowing systems and to
manage the subtleties of multi-threading with X11 in our code, it is
therefore preferable to retrieve this information before launching the
asynchronous image loading.

Fixes #76, related to !34.