Optimization: Switch to GQueue in RsttoImageList
This is for quick access to the last image in the list and the number of
images. The simple switch from GList to GQueue allows to decrease the
complexity of some operations from O(n) to O(1), while others are
slightly modified to take advantage of the new data structure when
possible.