Optimization: Use TreeModel IFace to walk the image list
Although the previous commit already reduces the complexity of browsing
the image list via an RsttoImageListIter to O(n), it is better to use
the GtkTreeModel interface in such cases. GtkTreeIter is indeed not
intended to be persistent, so its handling is simpler and faster:
allocation on the stack, no signal emission, and no addition to an iter
list to maintain. The stop condition at the end of the list is also
simpler.