Optimization: Add index and list link to RsttoImageListIter
This reduces the complexity of some operations from O(n) to O(1), in
particular `get_position()`, `iter_next()` and `iter_previous`. As this
iter is persistent and must therefore already be kept up to date when
the image list changes, this does not add any new maintenance constraint
(and thus potentially a new source of bugs), but only two more
attributes to be updated.