Thumbnailer: Remove upscale in case of unsupported flavor
While this is inconsequential in the current state of the code (the
maximum thumbnail size is less than `RSTTO_THUMBNAIL_FLAVOR_LARGE`),
this fallback is implemented in the wrong place.

Furthermore, it is actually not clear that it should be implemented,
because by the time support for larger sizes is added, an explicit
dependency on Xfce 4.18 could also be added (e.g. libxfce4util 4.18), so
that support for "x-large" and "xx-large" flavors could be assumed as a
prerequisite.

Amends previous commit.