Flatpak: Add Tumbler dependency as a private service
This makes the Flatpak application self-contained from a thumbnail
perspective, as it was already from a configuration perspective. So
there is no need to worry about whether we are in the Flatpak context
when building the thumbnail path: the cache directory is always
`$XDG_CACHE_HOME`, which varies depending on the context.

Related to xfce/tumbler!21, see !35 for more details.