Do not keep sticky state when current file is removed
Unlike what was done in d135e3ad185c90cf3527b786b856e1bfffe30cfc. This
allows in particular to stay on the current image when renaming it,
which is done via a deleted-created pair, at least when the monitoring
is done at the directory level. The case of file monitoring, which only
works halfway, is treated in the next commit.