thumbnailer: Remove invalid files after processing thumbnail queue
Removing a file from the image list usually causes its iterators to be
updated, which can then trigger a thumbnail request. Since queuing this
request modifies the internal thumbnailer list, you should not send a
remove request while browsing this list.

Given the thumbnail handling in Ristretto, unless there is a bug as
reported in #94, removing invalid files from the image list from the
thumbnailer should not cause a new thumbnail request. But it is safer to
move this removing after the GSourceFunc, i.e. into a GDestroyNotify.

Related: !31, #94