main_window.c: Use a better-known bus name
`org.xfce.FileManager` is specific to Thunar.  Without Thunar, Ristretto's
start-up is delayed for up to 25 seconds.

`org.freedesktop.FileManager1` on the other hand is a well-known name, which
is owned by many file managers, including Thunar.

This eliminates the delay and makes Ristretto agnostic.

Thanks to Daniel Shahaf for reporting and debugging.

Close #95.