Manage thumbs of thumbs correctly
As the specification states (there's not much choice) if the file is
itself a thumbnail, it should be used directly as a thumbnail. It seems
preferable to let tumbler determine whether the file is a thumbnail and
return the error signal provided for this purpose in its specification,
rather than determining ourselves whether the file is a thumbnail, with
a code that might not give quite the same result (and that might change
in tumbler).

Related: xfce/tumbler#94