clang-format: Add file
The original format is more or less preserved, except that
ContinuationIndentWidth has been changed to 4 instead of 8, and we break
a little less often at the first argument of functions (and never at the
first parameter of function definitions). We also try to be a little
more compact overall.

The space after the negation operator '!', which I myself had introduced
in the past, is dropped (I've changed my mind, it was definitely a bad
idea, sorry).

Related: apps/orage#32, xfce/xfce4-dev-tools!101